Better project ranking algorithm
The new system has been inductively shown to ``do the right thing`` 98%
of the time.  This comes with a caveat that ``the right thing`` is
a meaningless assertion.

It will always return A < B when s_A < s_B and b_A < b_B.  It will bias
ties in the sum of s_A and s_B toward the source code slightly.  It also
will bias the age of a project to favour a balance of updates (i.e. s -
b ~ 0).  A project that has recently updated one component, but has let
the other slip for two weeks will rank lower than a project that has
updated both in the past week.